refactor: like, user 통일
