+++ src/services/api/common.ts
@@ -1,0 +1,7 @@
+import { UserData, LikeData } from './types/common';
+
+export interface CommonService {
+  requestLogin(code: string): Promise<string | undefined>;
+  getUserInfo(accessToken: string | null): Promise<UserData>;
+  postLikeData(newsId: number): Promise<LikeData>;
+}
