--- src/services/api/like.ts
@@ -1,5 +1,0 @@
-import { LikeData } from './types/like';
-
-export interface LikeService {
-  postLikeData(newsId: number): Promise<LikeData>;
-}
