+++ src/services/api/types/common.ts
@@ -2,3 +2,8 @@ export interface UserData {
   nickname: string;
   email?: string;
 }
+
+export type LikeData = {
+  id: number;
+  isFavorite: boolean;
+};
