--- src/services/api/types/like.ts
@@ -1,4 +1,0 @@
-export type LikeData = {
-  id: number;
-  isFavorite: boolean;
-};
