+++ src/services/remote/common.ts
@@ -1,7 +1,8 @@
+import { CommonService } from '@src/services/api/common';
 import axios from 'axios';
-import { LoginUserService } from '../api/login-user';
+import { API } from './base';
 
-export function loginUserRemote(): LoginUserService {
+export function commonDataRemote(): CommonService {
   const requestLogin = async (code: string) => {
     try {
       const response = await axios.post(`https://deliverble.online/auth/authentication/kakao?code=${code}`);
@@ -14,9 +15,7 @@ export function loginUserRemote(): LoginUserService {
   const getUserInfo = async (accessToken: string | null) => {
     try {
       const response = await axios.get('https://deliverble.online/user', {
-        headers: {
-          Authorization: `Bearer ${accessToken}`,
-        },
+        headers: { Authorization: `Bearer ${accessToken}` },
       });
       return {
         nickname: response.data.data.nickname,
@@ -27,5 +26,15 @@ export function loginUserRemote(): LoginUserService {
     }
   };
 
-  return { requestLogin, getUserInfo };
+  const postLikeData = async (newsId: number) => {
+    const response = await API.post({ url: `/user/favorite/${newsId}` });
+    if (response.statusCode === 200) {
+      return {
+        id: response.data.newsId,
+        isFavorite: response.data.isFavorite,
+      };
+    } else throw '서버 통신 실패';
+  };
+
+  return { requestLogin, getUserInfo, postLikeData };
 }
